More integration test implementation
