+++ deeplearning4j/deeplearning4j-scaleout/deeplearning4j-scaleout-parallelwrapper-parameter-server/src/test/resources/log4j.properties
@@ -8,7 +8,6 @@ log4j.appender.Console.layout.ConversionPattern=%d{ABSOLUTE} %-5p ~ %m%n
 log4j.appender.org.springframework=DEBUG
 log4j.appender.org.nd4j=INFO
 log4j.logger.org.nd4j.aeron.ipc=INFO
-log4j.appender.org.canova=INFO
 log4j.appender.org.deeplearning4j=INFO
 log4j.appender.opennlp.uima=OFF
 log4j.appender.org.apache.uima=OFF
