+++ deeplearning4j/dl4j-integration-tests/pom.xml
@@ -75,6 +75,25 @@
                     <target>1.8</target>
                 </configuration>
             </plugin>
+
+            <!-- Skip this module for installation -->
+            <plugin>
+                <artifactId>maven-install-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>default-install</id>
+                        <phase>none</phase>
+                    </execution>
+                </executions>
+            </plugin>
+
+            <!-- Skip this module for deployment -->
+            <plugin>
+                <artifactId>maven-deploy-plugin</artifactId>
+                <configuration>
+                    <skip>true</skip>
+                </configuration>
+            </plugin>
         </plugins>
     </build>
